requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
peft>=0.6.0
trl>=0.7.0
datasets>=2.14.0
evaluate>=0.4.0

# Medical Imaging
pydicom>=2.4.0
nibabel>=5.1.0
SimpleITK>=2.3.0
opencv-python>=4.8.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API & Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# Speech Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
isort>=5.12.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.26.0
ipywidgets>=8.1.0

# Deployment
gunicorn>=21.2.0
docker>=6.1.0

# Model Optimization
onnx>=1.15.0
onnxruntime>=1.16.0
